requirements.txt
# pip install -r requirements.txt
requests
pulsar-client
python-dotenv
Flask
streamlit
astrapy
langgraph
langchain-openai
langchain
langchain-core
instructor